pyproject.toml
[project]
name = "appodeal-challenge"
version = "0.1.0"
description = "Appodeal Data Engineering Challenge"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.12.5",
    "pyspark==3.5.5",
    "pyyaml>=6.0.3",
]

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.9",
]
